- moved controllers to controller package - removed unused screen - added check for name and phone number in comment on first controller
