+++ src/be/mcjava/controller/CustomerScreenController.java
@@ -5,29 +5,38 @@ import javafx.fxml.FXML;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.Parent;
 import javafx.scene.Scene;
+import javafx.scene.control.TextField;
 import javafx.stage.Stage;
 
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
 public class CustomerScreenController {
-    //checks if it is a string with only letters.
-    public boolean isName(String name){
-        return name.chars().allMatch(Character::isLetter);
-
-    }
-    //checks valid phone-number
-    public boolean isPhoneNumber(String num){
-        Pattern p = Pattern.compile("(0/91)?[7-9][0-9]{9}");
-        Matcher m = p.matcher(num);
-        return (m.find() && m.group().equals(num));
-    }
     @FXML
-    private void continueFromLoginToMenuPressed(ActionEvent event) throws Exception{
-     FXMLLoader fxmlLoader = new FXMLLoader(getClass().getResource("../view/CustomerMainMenuOverview.fxml"));
-     Parent root1 = fxmlLoader.load();
-     Stage stage = new Stage();
-     stage.setScene(new Scene(root1));
-     stage.show();
+    TextField customername;
+
+    @FXML
+    TextField customerphonenumber;
+
+    //checks if it is a string with only letters.
+    public boolean isName(String name) {
+        return name.chars().allMatch(Character::isLetter);
+    }
+
+    //checks valid phone-number
+    public boolean isPhoneNumber(String customerPhoneNumber) {
+        Pattern p = Pattern.compile("(0/91)?[7-9][0-9]{9}");
+        Matcher m = p.matcher(customerPhoneNumber);
+        return (m.find() && m.group().equals(customerPhoneNumber));
+    }
+
+    @FXML
+    private void continueFromLoginToMenuPressed(ActionEvent event) throws Exception {
+        //if(isName(customername.getText()) && isPhoneNumber(customerphonenumber.getText())) {
+        FXMLLoader fxmlLoader = new FXMLLoader(getClass().getResource("../view/CustomerMainMenuOverview.fxml"));
+        Parent root1 = fxmlLoader.load();
+        Stage stage = new Stage();
+        stage.setScene(new Scene(root1));
+        stage.show();
     }
 }
