+++ src/be/mcjava/view/CustomerMainMenuOverview.fxml
@@ -10,7 +10,7 @@
 <?import javafx.scene.control.Label?>
 <FlowPane xmlns="http://javafx.com/javafx"
           xmlns:fx="http://javafx.com/fxml"
-          fx:controller="be.mcjava.view.MenuActionController"
+          fx:controller="be.mcjava.controller.MenuActionController"
           fx:id="flowpane"
           prefHeight="400.0" prefWidth="600.0">
     <stylesheets>
