+++ src/be/mcjava/view/CustomerMenuIngredientsChoice.fxml
@@ -10,7 +10,7 @@
 <?import javafx.scene.control.Label?>
 <FlowPane xmlns="http://javafx.com/javafx"
           xmlns:fx="http://javafx.com/fxml"
-          fx:controller="be.mcjava.view.MenuIngredientsActionController"
+          fx:controller="be.mcjava.controller.MenuIngredientsActionController"
           fx:id="flowpane"
           prefHeight="400.0" prefWidth="600.0">
     <stylesheets>
