--- src/be/mcjava/view/CustomerScreen2.fxml
@@ -1,10 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.scene.layout.*?>
-
-<AnchorPane xmlns="http://javafx.com/javafx"
-            xmlns:fx="http://javafx.com/fxml"
-            fx:controller="be.mcjava.controller.CustomerScreenController"
-            prefHeight="400.0" prefWidth="600.0">
-
-</AnchorPane>
